+++ sdk/core/azure-core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.0.0-beta.7 (Unreleased)
 
+### Breaking Changes
+
+- Split `Azure::Core::RequestConditions` into `Azure::Core::MatchConditions` and `Azure::Core::ModifiedConditions`.
+- Removed `TransportKind` enum from `Azure::Core::Http`.
 
 ## 1.0.0-beta.6 (2021-02-09)